Created map for demo 1
Created the main map for demo1. This includes the props, blockout and
dressing. Also includes game elements like the buttons and moving
platforms.

Added an outline shader for the demo1 map.

Props can be bounced and some even instance leaf particles.

Boars also react with hearts when they have been fed.

The collisions for the oak and acorn have been adjusted.

Warning: Windows added some files with uppercase letters and they cannot
be changed in Windows. This might mean that if you change branch or try
to merge you will get an error. If these issues persist, delete the
project and clone it again.