Finalize Dialogue functionality with trigger+sound
Added a DialogueTrigger node that receives an Area3D and a
DialogueContainer as paramenters. It starts the Dialogue when a
character enters the Area3D.

Wrote a Readme explaining how to use the Dialogue System with the
DialogueTrigger node.

Added sounds to advancing in the dialog and finishing it.

Fixed a problem with the avatar camera. It was showing ALL objects in
the scene, even the platforms, trees and such. Made it so it only shows
objects in layer 10 (which was renamed to "Dialogue")