Fix character visible in background
Initially I set up the Dialogue System to show anything in layer 1. This
is a problem because the level is also there, and it appeared in the
Dialogue UI.

To solve this, the character was set to Layer 10 (renamed as
"Dialogue"). This worked mostly ok, but it the player character, the one
moving on the level, was visible from the Dialogue UI's camera, it would
show up in the UI as well.

This should be the last fix. Instead of making the character visible in
layer 10 *always*, we only do it so when adding it to the Dialogue UI.