Fix FOR REAL THIS TIME input remapper smoke test
The last error was that the `default_input.data` file was being copied
as `default_input.data`, not the expected `input.data`. My bad. It's
been fixed now, so no more error messages from the smoke test!

Also the path was wrong, it should use `$HOME`, not the workspace
variable.