Fix 2 bugs: camera rotation & player movement
The first bug was that the camera was using some complex math to
rotate it. It made things more awkward and error prone. It has been
changed to a method using 2 empty Node3D to handle horizontal and
vertical rotation separately. Thanks for the tip Antimundo!

The second bug was that the player movement was behaving strangely
depending on where the camera was looking. This is because we were using
a secondary basis from the complex math aforementioned (weird stuff I
don't entirely understand). Since with the new implementation we don't
have that, we had to rework it. This new method also gives us the
horizontal rotation in a single node, so we can use that to rotate the
input vector easily. Everyone wins!