Fix integration test pipeline
Since the update to GUT's latest version the way errors are reported
changed and the script was not finding the errors and adding them up
correctly. That's been fixed.

Another thing from the update, the tests don't report
warnings anymore (they are reported as errors, which is good!).
Therefore, the warning detection has been removed.